requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
archspec @ file:///home/conda/feedstock_root/build_artifacts/archspec_1699370045702/work
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
boltons @ file:///home/conda/feedstock_root/build_artifacts/boltons_1703154663129/work
branca==0.7.1
Brotli @ file:///D:/bld/brotli-split_1695989908365/work
cachetools==5.3.3
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1700303426725/work/certifi
cffi @ file:///D:/bld/cffi_1696001730392/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1698833585322/work
click==8.1.7
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
conda @ file:///D:/bld/conda_1701731694793/work
conda-libmamba-solver @ file:///home/conda/feedstock_root/build_artifacts/conda-libmamba-solver_1702406360642/work/src
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1691048088238/work
conda_package_streaming @ file:///home/conda/feedstock_root/build_artifacts/conda-package-streaming_1691009212940/work
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.4
distro @ file:///home/conda/feedstock_root/build_artifacts/distro_1675116244235/work
Django==5.0.2
exceptiongroup==1.2.0
finance-datareader==0.9.70
flatbuffers==23.5.26
folium==0.16.0
fonttools==4.49.0
frozendict==2.4.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.4
httpx==0.27.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1701026962277/work
importlib-metadata==7.0.1
jax==0.4.25
Jinja2==3.1.3
jsonpatch @ file:///home/conda/feedstock_root/build_artifacts/jsonpatch_1695536281965/work
jsonpointer @ file:///D:/bld/jsonpointer_1695397563781/work
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.31
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langsmith==0.1.24
libmambapy @ file:///D:/bld/mamba-split_1702310523323/work/libmambapy
lxml==5.1.0
mamba @ file:///D:/bld/mamba-split_1702310523323/work/mamba
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.3
mdurl==0.1.2
mediapipe==0.10.10
menuinst @ file:///D:/bld/menuinst_1702317261150/work
ml-dtypes==0.3.2
multidict==6.0.5
multitasking==0.0.11
mypy-extensions==1.0.0
numpy==1.25.0
openai==1.13.3
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
orjson==3.9.15
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1696202382185/work
pandas==2.2.1
peewee==3.17.1
pillow==10.2.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1701708255999/work
pluggy @ file:///home/conda/feedstock_root/build_artifacts/pluggy_1693086607691/work
protobuf==3.20.3
pyarrow==15.0.0
pycosat @ file:///D:/bld/pycosat_1696356014994/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
PySocks @ file:///D:/bld/pysocks_1661604991356/work
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
requests-file==2.0.0
rich==13.7.1
rpds-py==0.18.0
ruamel.yaml @ file:///D:/bld/ruamel.yaml_1699007429890/work
ruamel.yaml.clib @ file:///D:/bld/ruamel.yaml.clib_1695997037989/work
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.4.6
soupsieve==2.5
SQLAlchemy==2.0.28
sqlparse==0.4.4
streamlit==1.31.1
streamlit-folium==0.18.0
tenacity==8.2.3
tiktoken==0.6.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1691671248568/work
truststore @ file:///home/conda/feedstock_root/build_artifacts/truststore_1694154605758/work
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1699933488691/work
validators==0.22.0
watchdog==4.0.0
webencodings==0.5.1
win-inet-pton @ file:///D:/bld/win_inet_pton_1667051142467/work
xyzservices==2023.10.1
yarl==1.9.4
yfinance==0.2.37
zipp==3.17.0
zstandard==0.22.0
